Rivisión de LoginActivity
Se han eliminado archivos y lineas de código inecesarios, y se ha retocado ligeramente la interfaz.
