--- Carlos_Proyecto_AndroidStudio/app/src/main/java/diazhernan/carlos/socialsports/data/model/LoggedInUser.java
@@ -1,23 +1,0 @@
-package diazhernan.carlos.socialsports.data.model;
-
-/**
- * Data class that captures user information for logged in users retrieved from LoginRepository
- */
-public class LoggedInUser {
-
-    private String userId;
-    private String displayName;
-
-    public LoggedInUser(String userId, String displayName) {
-        this.userId = userId;
-        this.displayName = displayName;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-}
