--- Carlos_Proyecto_AndroidStudio/app/src/main/java/diazhernan/carlos/socialsports/ui/login/LoginViewModelFactory.java
@@ -1,26 +1,0 @@
-package diazhernan.carlos.socialsports.ui.login;
-
-import androidx.lifecycle.ViewModel;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.annotation.NonNull;
-
-import diazhernan.carlos.socialsports.data.LoginDataSource;
-import diazhernan.carlos.socialsports.data.LoginRepository;
-
-/**
- * ViewModel provider factory to instantiate LoginViewModel.
- * Required given LoginViewModel has a non-empty constructor
- */
-public class LoginViewModelFactory implements ViewModelProvider.Factory {
-
-    @NonNull
-    @Override
-    @SuppressWarnings("unchecked")
-    public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if (modelClass.isAssignableFrom(LoginViewModel.class)) {
-            return (T) new LoginViewModel(LoginRepository.getInstance(new LoginDataSource()));
-        } else {
-            throw new IllegalArgumentException("Unknown ViewModel class");
-        }
-    }
-}
