+++ Carlos_Proyecto_AndroidStudio/app/src/main/res/values/strings.xml
@@ -6,10 +6,6 @@
     <string name="prompt_password">Password</string>
     <string name="action_sign_in">Sign in or register</string>
     <string name="action_sign_in_short">Sign in</string>
-    <string name="welcome">"Welcome !"</string>
-    <string name="invalid_username">Not a valid username</string>
-    <string name="invalid_password">Password must be >5 characters</string>
-    <string name="login_failed">"Login failed"</string>
 
     <!-- Strings related to main menu -->
     <string name="nav_new_event">Create New Ev.</string>
